Move preparing the rom to the rom module
